+++ crates/polars-plan/src/dsl/meta.rs
@@ -2,7 +2,7 @@ use std::fmt::Display;
 use std::ops::BitAnd;
 
 use super::*;
-use crate::logical_plan::projection::is_regex_projection;
+use crate::logical_plan::expr_expansion::is_regex_projection;
 use crate::logical_plan::tree_format::TreeFmtVisitor;
 use crate::logical_plan::visitor::{AexprNode, TreeWalker};
